created functioning app start page, need to create other functioning pages
