+++ projectshell/client/projectapp/src/Components/DeleteBtn/style.css
@@ -1,0 +1,5 @@
+.delete-btn {
+    float: right;
+    color: #e74944;
+  }
+
