+++ projectshell/client/projectapp/src/pages/watchList.js
@@ -1,0 +1,17 @@
+import React, { useState, useEffect } from "react";
+import DeleteBtn from "../Components";
+import Navbar from "..Components/Navbar";
+
+function watchList(props) {
+
+
+    return (
+        <>
+        <Navbar />
+        </>
+
+    );
+
+}
+
+export default watchList;
